feat: implement product and user seeding functions
